+++ flight-service/src/main/java/com/flightService/repository/TicketRepository.java
@@ -8,13 +8,13 @@ import java.util.List;
 public interface TicketRepository extends GenericJPARepository<Ticket> {
 
     @Query("from Ticket t LEFT JOIN FETCH t.flight f where f.id= ?1")
-    List<Ticket> findTicketsByFlightId(Long flightId);
+    List<Ticket> findTicketsByFlightId(long flightId);
 
     @Query("from Ticket t LEFT JOIN FETCH t.flight f where f.flightNumber= ?1")
-    List<Ticket> findTicketsByFlightNumber(int flightNumber);
+    List<Ticket> findTicketsByFlightNumber(long flightNumber);
 
     @Query("from Ticket t LEFT JOIN FETCH t.passenger p where p.id= ?1")
-    List<Ticket> findTicketsByPassengerId(Long passengerId);
+    List<Ticket> findTicketsByPassengerId(long passengerId);
 
     List<Ticket> findByActiveTrue();
